Bug 38428: Use links for stock rotation action buttons
This patch converts the action buttons on the stock rotation rota
manage items page to be links to avoid repeating form values (CSRF tokens, etc)

It moves the confirmation for removal from a rota to be a confirm and not a full page reload.

It also moves the fetch of a transfer into a shared variable.

Lastly, there is some chomping added to remove extra whitespace.

To test:
1 - Enable SotckRotation
2 - Create a rota
3 - Add some stages
4 - Add items to the rota
5 - Test the buttons to advance stage, add 'in demand' and remove
6 - Apply patch
7 - Confirm all buttons continue to function
8 - Sign off

Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>
Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>