Bug 36640: Fix display of 'Clear filter'
This is a weird one. When using KohaTable the first button was not
displayed if it was a custom button.
Doing the following fixed the problem (adding an empty first elt):
  dt_parameters["buttons"] = [{}]
  dt_parameters["buttons"].push({the_clear_filter_button});

However adding the definition of the button to $.fn.dataTable.ext.buttons fixed the problem.
As described on https://datatables.net/extensions/buttons/custom#Custom-button-type

This will definitely need to be improved later.(see bug 26553)

Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>