Bug 38462: Remove unused code for pagination in OPAC authority search
The remote branch for bug 13618 started out when OPAC authority search did its
own pagination in a template block, then bug 2735 replaced that with
pagination_bar, the template var passed from the script, then at some point
before bug 13618 landed the second time, a rebase error put the template block
back in. And since it's not at all obvious it is dead code unless you look for
where "pagination" rather than "pagination_bar" is used, we've kept updating
it, and bug 36032 thought the way it used >> instead of > for next ought to be
fixed while one that was actually used was fixed.

Time to get rid of it, before we 'fix' it again.

Test plan:
 1. Without the patch, in the OPAC go to Authority Search and do a search that
    will return pages of results, like the letter a
 2. Note that the button for the next page uses ">" rather than the "Next >>"
    in the block this patch removes
 3. Apply patch, click the browser Back button, refresh page
 4. Verify that when you search for something that returns more than one
    page of results, the page navigation hasn't changed from what it was
    before the patch, telling you that what the patch removes was unused

Sponsored-by: Chetco Community Public Library
Signed-off-by: David Nind <david@davidnind.com>
Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>