Bug 38485: Take user's previous change into consideration for holdings tables col vis
A lot happened, hard to summarize everything.

Basically at this point, we have (among other things):
* fixed bug 38118 (but reintroduced it with bug 36640)
* fixed bug 38436 (the original report was that the is_hidden setting
  was no longer taken into effect on the holdings tables)

Here we are trying to deal with making user's previous changes
persistent across the pagination and col vis changes.

Test plan:
Confirm that hiding checkouts in settings hide the columns
Confirm that showing checkouts then change page (*pagination*, so you need
more than 20 items) and confirm that the columns are still the same
Add a callnumber for an item from page 2 and confirm that the column is
hidden on page 1 and shown on page 2
Hide callnumber from the settings and confirm that the column is hidden
on pages 1 and 2
Show callnumber from the holdings page and confirm that the column is
displayed on page 2, go page 1: hidden, go back page 2: displayed.

And certainly way more other things to test, just confirm that the
behaviours with this patch are consistent with what a end user would
expect.

Note for QA/self: this is not very nice as column-visibility.dt event
can change in datatables.js and not impact this one.
Also the following nice sounds good and should be added to datatables.js
+            if (recalc === false) return;

Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>