Bug 38436: Fix OPAC tables
This patch sync columns_settings.inc for the OPAC-side

It however removes 'api' from datatables.js as it is never used.
'api' was renamed 'kohaTable' on bug 29408 but OPAC was forgotten.

Better to not have to deal with this unused code for now, we will
reintroduce it if needed later.

Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>