Bug 7508: (follow-up) Show item number if item has no barcode
This follow-up adds item number as a fallback for display when an item
doesn't have a barcode.

The patch also implements a sidebar sub-menu showing barcodes (or item
numbers) so that the user can jump directly to a particular section. The
sub-menu will show a scrollbar if the record has more than around 15
items.

The patch modifies staff-global.scss to improve readability of the
sidebar sub-menu and removes CSS which was specific to the system
preferences page. This make the sub-menus on the item details and system
preferences pages consistent with each other.

Signed-off-by: David Nind <david@davidnind.com>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>
Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>