Bug 38595: Fix broken table settings behaviour on some tables at the OPAC
Problem highlighted by bug 38436. There is a JS error on some views
with tables at the OPAC.

  Uncaught TypeError: table_settings.columns is undefined

Test plan:
Go to the 3 views modified by this patch and confirm that the JS error
in the console is gone when this patch is applied.

Signed-off-by: Owen Leonard <oleonard@myacpl.org>
Signed-off-by: Lucas Gass <lucas@bywatersolutions.com>
Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>
(cherry picked from commit 4733ca89a64039cf6084fc5d4f8905ecbce29955)
Signed-off-by: Paul Derscheid <paul.derscheid@lmscloud.de>