Bug 38620: Remove obsolete reference to hc-sicky on OPAC tags page
This patch removes code intended to include the hc-sticky js library.
hc-sticky is not longer included in Koha.

To test, apply the patch and make sure the TagsEnabled system preference
is turned on.

Log in to the OPAC and go to the tags page. Everything should work
correctly. There should be no errors in the developer console.

Signed-off-by: Lucas Gass <lucas@bywatersolutions.com>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>
Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>