Bug 38657: Prevent cover image from being from being obscured by the search results toolbar
On the OPAC search results page we're including the Chocolat image
viewer CSS after our global CSS file. This means that rules which
are intended to override the default Chocolat CSS aren't working. The
fix is to move the line which includes the Chocolat CSS.

To test, apply the patch and enable the LocalCoverImages and
OPACLocalCoverImages system preferences.

- If necessary, add a local cover image to a bibliographic record.
- Go to the OPAC and perform a search which will return multiple
  record including the record with the cover image.
- Click the cover image to trigger the image viewer.
- The image viewer should not appear to be "behind" the search results
  toolbar.

Sponsored-by: Athens County Public Libraries
Signed-off-by: David Nind <david@davidnind.com>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>
Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>