Bug 38729: Linker should consider diacritics
Currently, Koha linker does not pay attention to the diacritics while
linking--the decision about linking is taken based on the results of
search in the search engine (be it Zebra or Elasticsearch) which are
diacritics insensitive.  As a consequence, "Brückner, Karl" could be
erroneously linked to "Bruckner, Karl", in case when "Brückner, Karl"
did not exist in the authority file, or no link would be made in case
when both "Brückner, Karl" and "Bruckner, Karl" existed in the authority
file and LinkerModule is set to 'default'.

Test plan:
==========

1. Make sure that LinkerModule system preference is set to default.
2. Create an authority record of type PERSO_NAME with a name containing
   diacritics (e.g. "Brückner, Karl" , or "Rąkowski, Albin" etc.).
3. Edit an existing or create a new bibliographic record and put in a
   personal name field (100, 600, 700) that name without diacritics.
4. Click "Link authorities automatically" -- the two different names will
   be linked, which is wrong.
5. Create a second authority record with name form exactly from the
   bibliographic record (without diacritics).
6. Edit an existing or create a new bibliographic record and put in a
   personal name field (100, 600, 700) the name without diacritics.
7. Click "Link authorities automatically" -- you should get a message
   "More than one local match found. Possibly a duplicate authority!".
8. Apply the patch ; restart_all.
9. Choose another name pair with and without diacritics. Repeat p. 3-4.
   The bibliographic field should not be linked this time.
10. With the new pair, repeat p. 5.-7.  The bibliographic field should
    now be correctly linked.

Sponsored-by: Ignatianum University in Cracow
Signed-off-by: Roman Dolny <roman.dolny@jezuici.pl>